requirements.txt
Flask
gunicorn
python-dotenv
boto3